Split worker runtime tests by responsibility
