+++ Tests/SpeakSwiftlyTests/Runtime/WorkerRuntimePlaybackTests.swift
@@ -1,0 +1,159 @@
+import Foundation
+import Testing
+@testable import SpeakSwiftlyCore
+import TextForSpeech
+
+// MARK: - Live Playback Queueing
+
+@Test func speakLiveBackgroundAcknowledgesQueueBeforePlaybackStartsAndOnlySucceedsOnce() async throws {
+    let output = OutputRecorder()
+    let playbackDrain = AsyncGate()
+    let playback = PlaybackSpy(behavior: .gate(playbackDrain))
+    let storeRoot = makeTempDirectoryURL()
+    defer { try? FileManager.default.removeItem(at: storeRoot) }
+
+    let store = try makeProfileStore(rootURL: storeRoot)
+    _ = try store.createProfile(
+        profileName: "default-femme",
+        modelRepo: "test-model",
+        voiceDescription: "Warm and bright.",
+        sourceText: "Reference transcript",
+        sampleRate: 24_000,
+        canonicalAudioData: Data([0x01, 0x02])
+    )
+
+    let runtime = try await makeRuntime(
+        rootURL: storeRoot,
+        output: output,
+        playback: playback,
+        residentModelLoader: { _ in makeResidentModel() }
+    )
+
+    await runtime.start()
+    #expect(await waitUntil {
+        output.containsJSONObject {
+            $0["event"] as? String == "worker_status"
+                && $0["stage"] as? String == "resident_model_ready"
+        }
+    })
+
+    let activeID = await runtime.speak(
+        text: "Hello there",
+        with: "default-femme",
+        as: .live,
+        id: "req-1"
+    ).id
+    #expect(activeID == "req-1")
+    #expect(await waitUntil {
+        output.containsJSONObject {
+            $0["id"] as? String == "req-1"
+                && $0["event"] as? String == "progress"
+                && $0["stage"] as? String == "preroll_ready"
+        }
+    })
+
+    let backgroundID = await runtime.speak(
+        text: "Hi there",
+        with: "default-femme",
+        as: .live,
+        id: "req-2"
+    ).id
+    #expect(backgroundID == "req-2")
+
+    #expect(await waitUntil {
+        output.containsJSONObject {
+            $0["id"] as? String == "req-2"
+                && $0["ok"] as? Bool == true
+        }
+    })
+    #expect(await waitUntil {
+        output.containsJSONObject {
+            $0["id"] as? String == "req-2"
+                && $0["event"] as? String == "started"
+                && $0["op"] as? String == "queue_speech_live"
+        }
+    })
+    #expect(!output.containsJSONObject {
+        $0["id"] as? String == "req-2"
+            && $0["event"] as? String == "progress"
+            && $0["stage"] as? String == "playback_finished"
+    })
+
+    await playbackDrain.open()
+
+    #expect(await waitUntil {
+        output.containsJSONObject {
+            $0["id"] as? String == "req-2"
+                && $0["event"] as? String == "progress"
+                && $0["stage"] as? String == "playback_finished"
+        }
+    })
+    #expect(output.countJSONObjects {
+        $0["id"] as? String == "req-2"
+            && $0["ok"] as? Bool == true
+    } == 1)
+}
+
+@Test func speakLiveBackgroundCanFailAfterEnqueueAcknowledgement() async throws {
+    let output = OutputRecorder()
+    let storeRoot = makeTempDirectoryURL()
+    defer { try? FileManager.default.removeItem(at: storeRoot) }
+
+    let store = try makeProfileStore(rootURL: storeRoot)
+    _ = try store.createProfile(
+        profileName: "default-femme",
+        modelRepo: "test-model",
+        voiceDescription: "Warm and bright.",
+        sourceText: "Reference transcript",
+        sampleRate: 24_000,
+        canonicalAudioData: Data([0x01, 0x02])
+    )
+
+    let runtime = try await makeRuntime(
+        rootURL: storeRoot,
+        output: output,
+        playback: PlaybackSpy(
+            behavior: .throw(
+                WorkerError(
+                    code: .audioPlaybackFailed,
+                    message: "Background playback failed in the test playback controller after the request had already been accepted."
+                )
+            )
+        ),
+        residentModelLoader: { _ in makeResidentModel() }
+    )
+
+    await runtime.start()
+    #expect(await waitUntil {
+        output.containsJSONObject {
+            $0["event"] as? String == "worker_status"
+                && $0["stage"] as? String == "resident_model_ready"
+        }
+    })
+
+    _ = await runtime.speak(
+        text: "Hello there",
+        with: "default-femme",
+        as: .live,
+        id: "req-fail"
+    )
+
+    #expect(await waitUntil {
+        output.containsJSONObject {
+            $0["id"] as? String == "req-fail"
+                && $0["ok"] as? Bool == true
+        }
+    })
+    #expect(await waitUntil {
+        output.containsJSONObject {
+            $0["id"] as? String == "req-fail"
+                && $0["ok"] as? Bool == false
+                && $0["code"] as? String == "audio_playback_failed"
+        }
+    })
+    #expect(output.countJSONObjects {
+        $0["id"] as? String == "req-fail"
+            && $0["ok"] as? Bool == true
+    } == 1)
+}
+
