+++ Tests/SpeakSwiftlyTests/Runtime/WorkerRuntimeShutdownTests.swift
@@ -1,0 +1,318 @@
+import Foundation
+import Testing
+@testable import SpeakSwiftlyCore
+import TextForSpeech
+
+// MARK: - Shutdown Behavior
+
+@Test func shutdownCancelsActivePlaybackAndQueuedRequestsExactlyOnce() async throws {
+    let output = OutputRecorder()
+    let storeRoot = makeTempDirectoryURL()
+    defer { try? FileManager.default.removeItem(at: storeRoot) }
+
+    let store = try makeProfileStore(rootURL: storeRoot)
+    _ = try store.createProfile(
+        profileName: "default-femme",
+        modelRepo: "test-model",
+        voiceDescription: "Warm and bright.",
+        sourceText: "Reference transcript",
+        sampleRate: 24_000,
+        canonicalAudioData: Data([0x01, 0x02])
+    )
+
+    let playback = PlaybackSpy(behavior: .sleep(.seconds(30)))
+    let runtime = try await makeRuntime(
+        rootURL: storeRoot,
+        output: output,
+        playback: playback,
+        residentModelLoader: { _ in makeResidentModel() }
+    )
+
+    await runtime.start()
+    #expect(await waitUntil {
+        output.containsJSONObject {
+            $0["event"] as? String == "worker_status"
+                && $0["stage"] as? String == "resident_model_ready"
+        }
+    })
+
+    await runtime.accept(line: #"{"id":"req-1","op":"queue_speech_live","text":"Hello there","profile_name":"default-femme"}"#)
+    #expect(await waitUntil {
+        output.containsJSONObject {
+            $0["id"] as? String == "req-1"
+                && $0["event"] as? String == "progress"
+                && $0["stage"] as? String == "buffering_audio"
+        }
+    })
+    #expect(await waitUntil {
+        output.containsJSONObject {
+            $0["id"] as? String == "req-1"
+                && $0["event"] as? String == "progress"
+                && $0["stage"] as? String == "preroll_ready"
+        }
+    })
+
+    await runtime.accept(line: #"{"id":"req-2","op":"queue_speech_live","text":"Hello again","profile_name":"default-femme"}"#)
+    #expect(await waitUntil {
+        output.containsJSONObject {
+            $0["id"] as? String == "req-2"
+                && $0["ok"] as? Bool == true
+        }
+    })
+    #expect(await waitUntil {
+        output.containsJSONObject {
+            $0["id"] as? String == "req-2"
+                && $0["event"] as? String == "started"
+                && $0["op"] as? String == "queue_speech_live"
+        }
+    })
+
+    await runtime.shutdown()
+
+    #expect(await waitUntil {
+        output.containsJSONObject {
+            $0["id"] as? String == "req-1"
+                && $0["ok"] as? Bool == false
+                && $0["code"] as? String == "request_cancelled"
+        }
+    })
+    #expect(await waitUntil {
+        output.containsJSONObject {
+            $0["id"] as? String == "req-2"
+                && $0["ok"] as? Bool == false
+                && $0["code"] as? String == "request_cancelled"
+        }
+    })
+    #expect(output.countJSONObjects {
+        $0["id"] as? String == "req-1"
+            && $0["ok"] as? Bool == false
+    } == 1)
+    #expect(playback.stopCount == 1)
+}
+
+@Test func shutdownPathEmitsCancellationNotPlaybackTimeout() async throws {
+    let output = OutputRecorder()
+    let storeRoot = makeTempDirectoryURL()
+    defer { try? FileManager.default.removeItem(at: storeRoot) }
+
+    let store = try makeProfileStore(rootURL: storeRoot)
+    _ = try store.createProfile(
+        profileName: "default-femme",
+        modelRepo: "test-model",
+        voiceDescription: "Warm and bright.",
+        sourceText: "Reference transcript",
+        sampleRate: 24_000,
+        canonicalAudioData: Data([0x01, 0x02])
+    )
+
+    let playback = PlaybackSpy(behavior: .sleep(.seconds(30)))
+    let runtime = try await makeRuntime(
+        rootURL: storeRoot,
+        output: output,
+        playback: playback,
+        residentModelLoader: { _ in makeResidentModel() }
+    )
+
+    await runtime.start()
+    #expect(await waitUntil {
+        output.containsJSONObject {
+            $0["event"] as? String == "worker_status"
+                && $0["stage"] as? String == "resident_model_ready"
+        }
+    })
+
+    await runtime.accept(line: #"{"id":"req-1","op":"queue_speech_live","text":"Hello there","profile_name":"default-femme"}"#)
+    #expect(await waitUntil {
+        output.containsJSONObject {
+            $0["id"] as? String == "req-1"
+                && $0["event"] as? String == "progress"
+                && $0["stage"] as? String == "preroll_ready"
+        }
+    })
+
+    await runtime.shutdown()
+
+    #expect(await waitUntil {
+        output.containsJSONObject {
+            $0["id"] as? String == "req-1"
+                && $0["ok"] as? Bool == false
+                && $0["code"] as? String == "request_cancelled"
+        }
+    })
+    #expect(!output.containsJSONObject {
+        $0["id"] as? String == "req-1"
+            && $0["ok"] as? Bool == false
+            && $0["code"] as? String == "audio_playback_timeout"
+    })
+}
+
+@Test func shutdownFailsTypedRequestStreamsForActiveAndQueuedRequests() async throws {
+    let output = OutputRecorder()
+    let playbackDrain = AsyncGate()
+    let profileGate = AsyncGate()
+    let playback = PlaybackSpy(behavior: .gate(playbackDrain))
+    let storeRoot = makeTempDirectoryURL()
+    defer { try? FileManager.default.removeItem(at: storeRoot) }
+
+    let store = try makeProfileStore(rootURL: storeRoot)
+    _ = try store.createProfile(
+        profileName: "default-femme",
+        modelRepo: "test-model",
+        voiceDescription: "Warm and bright.",
+        sourceText: "Reference transcript",
+        sampleRate: 24_000,
+        canonicalAudioData: Data([0x01, 0x02])
+    )
+
+    let runtime = try await makeRuntime(
+        rootURL: storeRoot,
+        output: output,
+        playback: playback,
+        residentModelLoader: { _ in makeResidentModel() },
+        profileModelLoader: {
+            makeProfileModel {
+                await profileGate.wait()
+            }
+        }
+    )
+
+    await runtime.start()
+    #expect(await waitUntil {
+        output.containsJSONObject {
+            $0["event"] as? String == "worker_status"
+                && $0["stage"] as? String == "resident_model_ready"
+        }
+    })
+
+    let activeHandle = await runtime.submit(
+        .queueSpeech(
+            id: "req-active-shutdown-stream",
+            text: "Hello there",
+            profileName: "default-femme",
+            textProfileName: nil,
+            jobType: .live,
+            textContext: nil,
+            sourceFormat: nil
+        )
+    )
+    var activeIterator = activeHandle.events.makeAsyncIterator()
+
+    let activeStarted = try await activeIterator.next()
+    #expect(
+        activeStarted == .acknowledged(
+            WorkerSuccessResponse(id: "req-active-shutdown-stream")
+        )
+    )
+
+    let queuedHandle = await runtime.submit(
+        .createProfile(
+            id: "req-queued-shutdown-stream",
+            profileName: "shutdown-queued-profile",
+            text: "A queued request that should still be active when shutdown begins.",
+            vibe: .femme,
+            voiceDescription: "Warm and bright",
+            outputPath: nil
+        )
+    )
+    var queuedIterator = queuedHandle.events.makeAsyncIterator()
+
+    let queuedEvent = try await queuedIterator.next()
+    #expect(queuedEvent == .started(WorkerStartedEvent(id: "req-queued-shutdown-stream", op: "create_profile")))
+
+    await runtime.shutdown()
+
+    do {
+        while let _ = try await activeIterator.next() {}
+        Issue.record("The active typed request stream should have thrown during shutdown.")
+    } catch let error as WorkerError {
+        #expect(error.code == .requestCancelled)
+    }
+
+    do {
+        while let _ = try await queuedIterator.next() {}
+        Issue.record("The queued typed request stream should have thrown during shutdown.")
+    } catch let error as WorkerError {
+        #expect(error.code == .requestCancelled)
+    }
+
+    await profileGate.open()
+    await playbackDrain.open()
+}
+
+@Test func shutdownRejectsNewRequests() async throws {
+    let output = OutputRecorder()
+    let runtime = try await makeRuntime(
+        output: output,
+        playback: PlaybackSpy(),
+        residentModelLoader: { _ in makeResidentModel() }
+    )
+
+    await runtime.start()
+    await runtime.shutdown()
+    await runtime.accept(line: #"{"id":"req-1","op":"list_profiles"}"#)
+
+    #expect(await waitUntil {
+        output.containsJSONObject {
+            $0["id"] as? String == "req-1"
+                && $0["ok"] as? Bool == false
+                && $0["code"] as? String == "worker_shutting_down"
+        }
+    })
+}
+
+@Test func shutdownCancelsActiveProfileCreationBeforeProfileIsWritten() async throws {
+    let output = OutputRecorder()
+    let storeRoot = makeTempDirectoryURL()
+    defer { try? FileManager.default.removeItem(at: storeRoot) }
+
+    let runtime = try await makeRuntime(
+        rootURL: storeRoot,
+        output: output,
+        playback: PlaybackSpy(),
+        residentModelLoader: { _ in makeResidentModel() },
+        profileModelLoader: {
+            AnySpeechModel(
+                sampleRate: 24_000,
+                generate: { _, _, _, _, _, _ in
+                    try await Task.sleep(for: .seconds(30))
+                    return [0.1, 0.2, 0.3]
+                },
+                generateSamplesStream: { _, _, _, _, _, _, _ in
+                    AsyncThrowingStream { continuation in
+                        continuation.finish()
+                    }
+                }
+            )
+        }
+    )
+
+    await runtime.start()
+    #expect(await waitUntil {
+        output.containsJSONObject {
+            $0["event"] as? String == "worker_status"
+                && $0["stage"] as? String == "resident_model_ready"
+        }
+    })
+
+    await runtime.accept(
+        line: #"{"id":"req-1","op":"create_profile","profile_name":"bright-guide","text":"Hello there","vibe":"femme","voice_description":"Warm and bright"}"#
+    )
+    #expect(await waitUntil {
+        output.containsJSONObject {
+            $0["id"] as? String == "req-1"
+                && $0["event"] as? String == "progress"
+                && $0["stage"] as? String == "generating_profile_audio"
+        }
+    })
+
+    await runtime.shutdown()
+
+    #expect(await waitUntil {
+        output.containsJSONObject {
+            $0["id"] as? String == "req-1"
+                && $0["ok"] as? Bool == false
+                && $0["code"] as? String == "request_cancelled"
+        }
+    })
+    #expect(!FileManager.default.fileExists(atPath: storeRoot.appendingPathComponent("bright-guide").path))
+}
